@@ -13,11 +13,6 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^2.0.18",
-    "@chakra-ui/react": "^2.5.5",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@linaria/babel-preset": "^4.4.3",
     "@linaria/core": "^4.2.8",
     "@linaria/react": "^4.3.6",
     "@next/bundle-analyzer": "^13.3.0",
@@ -28,29 +23,34 @@
     "@types/react": "18.0.35",
     "@types/react-dom": "18.0.11",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "framer-motion": "^10.12.2",
     "graphql": "^16.6.0",
     "groq": "^3.8.3",
-    "next": "13.3.0",
-    "next-sanity": "^4.1.7",
+    "next": "^13.4.3",
+    "next-sanity": "^4.3.2",
     "next-with-linaria": "^0.5.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "require": "^0.4.4",
     "typescript": "5.0.4"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.0.5",
-    "@storybook/addon-interactions": "^7.0.5",
-    "@storybook/addon-links": "^7.0.5",
-    "@storybook/addon-mdx-gfm": "^7.0.5",
-    "@storybook/blocks": "^7.0.5",
-    "@storybook/nextjs": "^7.0.5",
-    "@storybook/react": "^7.0.5",
-    "@storybook/testing-library": "^0.0.14-next.2",
-    "eslint-plugin-storybook": "^0.6.11",
-    "storybook": "^7.0.5"
+    "@linaria/babel-preset": "^4.4.5",
+    "@linaria/webpack-loader": "^4.1.17",
+    "@storybook/addon-essentials": "^7.0.12",
+    "@storybook/addon-interactions": "^7.0.12",
+    "@storybook/addon-links": "^7.0.12",
+    "@storybook/addon-mdx-gfm": "^7.0.12",
+    "@storybook/blocks": "^7.0.12",
+    "@storybook/builder-webpack5": "^7.0.12",
+    "@storybook/manager-webpack5": "^6.5.16",
+    "@storybook/nextjs": "^7.0.12",
+    "@storybook/react": "^7.0.12",
+    "@storybook/testing-library": "^0.1.0",
+    "@svgr/webpack": "^8.0.1",
+    "eslint-config-next": "^13.4.2",
+    "eslint-plugin-storybook": "^0.6.12",
+    "mini-css-extract-plugin": "^2.7.5",
+    "storybook": "^7.0.12"
   }
 }
