@@ -5,32 +5,30 @@ import { css, cx } from '@linaria/core';
 import React, { AnchorHTMLAttributes, ButtonHTMLAttributes, PropsWithChildren } from 'react';
 
 // Write your styles in `css` tag
-const button = css`
+const label = css`
   text-transform: uppercase;
   color: #fff;
 	background-color: var(--c-primary);
-	padding: 10px 20px;
+	padding: var(--s-xxs) var(--s-s);
 	text-align: center;
 	border-radius: 5px;
 	cursor: pointer;
 	transition: all 0.3s ease-in-out;
 	border: none;
 	text-decoration: none;
+	font-size: var(--fs-xxs);
 
 	&:hover {
 		background-color: #000;
 	}
 `;
 
-type CtaButton = ButtonHTMLAttributes<HTMLButtonElement>;
-
-export default function CtaButton({ children, className, ...props }: PropsWithChildren<CtaButton>) {
+export default function Label({ children }: PropsWithChildren<{}>) {
 	return (
-		<button
-			className={cx(button, className)}
-			{...props}
+		<span
+			className={cx(label)}
 		>
 			{children}
-		</button>
+		</span>
 	);
 }
