@@ -5,7 +5,6 @@ import { getSanityImageUrl } from '@data'
 import { PortableText } from "@portabletext/react";
 import MovieGrid from '@components/MovieGrid';
 import PersonCard from '@components/PersonCard';
-import { Box, Heading, SimpleGrid } from '@chakra-ui/react';
 import Head from 'next/head';
 
 type Props = {
@@ -35,13 +34,9 @@ const MovieView = ({movie}: Props) => {
 			>
 				<PortableText value={movie.overview} />
 			</DetailView>
-			<SimpleGrid columns={1} spacing={5}>
-				<Box>
-					<Heading as='h2' size='2xl' textAlign={'center'}>
-						Characters
-					</Heading>
-				</Box>
-			</SimpleGrid>
+			<h2>
+				Characters
+			</h2>
 			{
 				movie.castMembers && (
 					<MovieGrid>
